// src/app/models/ISupermaestro.ts
export interface ISupermaestro {
  idSupermaestro: number;
  generico: string | null;
  bPautas: string | null;
  codigoPautas: string | null;
  bWip: string | null;
  codigoWip: string | null;
  bAlternativo1: string | null;
  alternativo1: string | null;
  bAlternativo2: string | null;
  alternativo2: string | null;
  puesto: string | null;
  gaveta: string | null;
  descripcion: string | null;
  descripSector: string | null;
  area: string | null;
  cantidadMaterial: string | null;
  stockGaveta: string | null;
  stockSeguridad: string | null;
  fecha: string | null;
}